refactor: ControllerFactory 작성, Controller 구현체 중복 코드 제거
