+++ src/main/java/blackjack/Main.java
@@ -1,12 +1,13 @@
 package blackjack;
 
 import blackjack.controller.BlackJackController;
+import blackjack.controller.ControllerFactory;
 import blackjack.view.InputView;
 
 public class Main {
     public static void main(String[] args) {
         String selectedGame = InputView.enterGame();
-        BlackJackController controller = BlackJackController.of(selectedGame);
-        controller.run();
+        BlackJackController controller = ControllerFactory.of(selectedGame);
+        controller.play();
     }
 }
