+++ src/main/java/blackjack/controller/BettingGame.java
@@ -1,6 +1,9 @@
 package blackjack.controller;
 
-import blackjack.domain.participant.*;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Money;
+import blackjack.domain.participant.Name;
+import blackjack.domain.participant.Players;
 import blackjack.domain.result.PlayerResult;
 import blackjack.domain.result.ResponseDTO.ProfitDTO;
 import blackjack.view.InputView;
@@ -13,18 +16,10 @@ import java.util.stream.Collectors;
 public class BettingGame extends BlackJackController {
 
     @Override
-    public void run() {
+    protected Players createPlayers() {
         List<Name> names = createNames();
         List<Money> bettingMoneys = createMoneys(names);
-
-        Players players = new Players(names, bettingMoneys);
-        Dealer dealer = new Dealer();
-
-        dealFirstCards(dealer, players);
-
-        dealAdditionalCards(players, dealer);
-
-        showProfit(players, dealer);
+        return new Players(names, bettingMoneys);
     }
 
     private List<Name> createNames() {
@@ -39,46 +34,8 @@ public class BettingGame extends BlackJackController {
                 .collect(Collectors.toList());
     }
 
-
-    private void dealFirstCards(Dealer dealer, Players players) {
-        deck.dealFirstCards(dealer);
-
-        for (Player player : players.getPlayers()) {
-            deck.dealFirstCards(player);
-        }
-
-        OutputView.printInitialStatus(players, dealer);
-    }
-
-    private void dealAdditionalCards(Players players, Dealer dealer) {
-        for (Player player : players.getPlayers()) {
-            dealAdditionalCards(player);
-        }
-
-        dealAdditionalCards(dealer);
-
-        OutputView.printFinalStatus(players, dealer);
-    }
-
-    private void dealAdditionalCards(Player player) {
-        while (player.canGetMoreCard() && player.wantMoreCard(readYesOrNo(player))) {
-            deck.dealAdditionalCard(player);
-            OutputView.printCardsStatus(player.name(), player.showCards());
-        }
-    }
-
-    private void dealAdditionalCards(Dealer dealer) {
-        while (dealer.canGetMoreCard()) {
-            deck.dealAdditionalCard(dealer);
-            OutputView.printDealerGetMoreCard();
-        }
-    }
-
-    private String readYesOrNo(Player player) {
-        return InputView.readYesOrNo(player.name());
-    }
-
-    private void showProfit(Players players, Dealer dealer) {
+    @Override
+    protected void showResult(Players players, Dealer dealer) {
         List<PlayerResult> playerResults = players.createPlayerResults(dealer);
 
         List<ProfitDTO> playerDTOS = playerResults.stream()
